Styling Changes, added backgroundcolor to main in header.css, moved Menucard.css and tsx to MenuCard folder in components folder, made small change to Menupage.tsx and styled ThaiMenuPage.
