+++ packages/baan-thai-frontend/src/components/Header/Header.css
@@ -8,6 +8,10 @@
 	padding: 0;
 }
 
+main {
+	background-color: #E4DACE;
+}
+
 .header {
 	display: flex;
 	flex-direction: row;
