--- packages/baan-thai-frontend/src/components/MenuCard.css
@@ -1,48 +1,0 @@
-
-/* Endast tillfälligt styling */
-.menu-card {
-  border: 1px solid #ddd;
-  border-radius: 12px;
-  padding: 16px;
-  width: 100%;
-  height: 100%;
-  min-height: 200px;
-  box-sizing: border-box;
-  background: rgb(29, 27, 27);
-  box-shadow: 0 2px 6px rgba(0,0,0,0.1);
-  display: flex;
-  flex-direction: column;
-}
-
-.menu-card h3 {
-  margin: 0 0 8px;
-  flex-shrink: 0;
-}
-
-.menu-card .description {
-  margin: 0 0 12px;
-  color: #555;
-  flex-grow: 1;
-}
-
-.menu-card .price {
-  font-weight: bold;
-  margin-bottom: 12px;
-}
-
-.menu-card .add-button {
-  padding: 8px 12px;
-  background: #ff9800;
-  color: rgb(163, 105, 105);
-  border: none;
-  border-radius: 8px;
-  cursor: pointer;
-  width: 100%;
-  transition: all 0.3s ease;
-}
-
-.menu-card .add-button:hover {
-  background: #f57c00;
-  transform: translateY(-2px);
-  box-shadow: 0 4px 8px rgba(255, 152, 0, 0.4);
-}
